requirements.txt
agent-framework>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
